# openapi/wsapi-api.yml
openapi: 3.0.3
info:
  title: WSAPI - WhatsApp API
  description: RESTful HTTP API for WhatsApp via whatsmeow. Supports multi-instance connections, webhook event delivery, and HMAC-SHA256 signing.
  version: 2.0.0
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
security:
  - InstanceAuth: []
paths:
...
  /messages/{messageId}/read:
    post:
      tags: [Messages]
      summary: Mark a message as read
      description: Mark a message as read with the specified receipt type.
      operationId: markAsRead
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/MessageIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MarkAsReadRequest"
            example:
              chatId: "1234567890@s.whatsapp.net"
              senderId: "1234567890@s.whatsapp.net"
              receiptType: "read"
      responses:
        "204":
          description: Message marked as read
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /users/me/presence:
    put:
      tags: [Users]
      summary: Set presence state
      description: Set the account's presence to available or unavailable. This is a transient action (not persisted in the profile).
      operationId: setPresence
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetMyPresenceRequest"
      responses:
        "204":
          description: Presence updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
...
  /chats:
    get:
      tags: [Chats]
      summary: List all known chats
      description: Returns all known chats for the instance, enriched with contact info (pushName, businessName, fullName) and chat settings (pinned, archived, muted). Sorted by last activity descending.
      operationId: listChats
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
      responses:
        "200":
          description: List of chats
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ChatListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}:
    get:
      tags: [Chats]
      summary: Get chat info
      description: Returns details for a single chat, enriched with contact info and chat settings.
      operationId: getChatInfo
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "200":
          description: Chat info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatListItem"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

    delete:
      tags: [Chats]
      summary: Delete a chat
      operationId: deleteChat
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "204":
          description: Chat deleted
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/picture:
    get:
      tags: [Chats]
      summary: Get chat profile picture
      operationId: getChatPicture
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "200":
          description: Chat picture info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChatPictureResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/business:
    get:
      tags: [Chats]
      summary: Get business profile for a chat
      operationId: getChatBusinessProfile
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "200":
          description: Business profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BusinessProfileResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/presence:
    put:
      tags: [Chats]
      summary: Send chat presence (typing, paused, recording)
      operationId: setChatPresence
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPresenceRequest"
      responses:
        "204":
          description: Presence sent
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/presence/subscribe:
    put:
      tags: [Chats]
      summary: Subscribe to presence updates for a chat
      operationId: subscribeChatPresence
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "204":
          description: Subscribed
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/ephemeral:
    put:
      tags: [Chats]
      summary: Set disappearing messages timer
      operationId: setChatEphemeral
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetChatEphemeralRequest"
      responses:
        "204":
          description: Timer set
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/mute:
    put:
      tags: [Chats]
      summary: Mute or unmute a chat
      operationId: muteChat
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MuteChatRequest"
      responses:
        "204":
          description: Mute setting updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/pin:
    put:
      tags: [Chats]
      summary: Pin or unpin a chat
      operationId: pinChat
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PinChatRequest"
      responses:
        "204":
          description: Pin setting updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/archive:
    put:
      tags: [Chats]
      summary: Archive or unarchive a chat
      operationId: archiveChat
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ArchiveChatRequest"
      responses:
        "204":
          description: Archive setting updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/read:
    put:
      tags: [Chats]
      summary: Mark a chat as read or unread
      operationId: markChatAsRead
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatMarkAsReadRequest"
      responses:
        "204":
          description: Read state updated
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/messages:
    post:
      tags: [Chats]
      summary: Request on-demand message history
      description: |
        Triggers an on-demand history sync request for the given chat. The response arrives
        asynchronously as a `message_history_sync` event via the configured publisher.
      operationId: requestChatMessages
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RequestMessagesRequest"
            example:
              lastMessageId: "3EB0A0B0C1D2E3F4"
              lastMessageSenderId: "1234567890"
              count: 50
      responses:
        "202":
          description: History sync request accepted
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"
  /chats/{chatId}/clear:
    post:
      tags: [Chats]
      summary: Clear all messages from a chat
      operationId: clearChat
      parameters:
        - $ref: "#/components/parameters/XInstanceId"
        - $ref: "#/components/parameters/ChatIdPath"
      responses:
        "204":
          description: Chat cleared
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "409":
          $ref: "#/components/responses/DeviceNotPaired"
        "503":
          $ref: "#/components/responses/ServiceUnavailable"

  # ─── Calls ─────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    AdminAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Admin API key for instance management endpoints (`/admin/instances/*`). Set via `auth.adminApiKey` config.

    InstanceAuth:
      type: apiKey
      in: header
      name: X-Api-Key
      description: Instance API key. Resolved from the instance's own `apiKey` or the global `auth.defaultApiKey`.
  parameters:
    XInstanceId:
      name: X-Instance-Id
      in: header
      required: true
      description: The ID of the WhatsApp instance to use
      schema:
        type: string
        default: "{{instanceId}}"
...
    MessageIdPath:
      name: messageId
      in: path
      required: true
      description: Message ID
      schema:
        type: string
...
    ChatIdPath:
      name: chatId
      in: path
      required: true
      description: Chat JID (user or group)
      schema:
        type: string
...
  responses:
...
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 400
            detail: "invalid request body"
    Unauthorized:
      description: Authentication failed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 401
            detail: "unauthorized"
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 404
            detail: "not found"
    ServiceUnavailable:
      description: Instance not available (not connected or service is nil)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 503
            detail: "instance not available"
    DeviceNotPaired:
      description: Device is not paired. Complete the pairing process first.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
          example:
            status: 403
            detail: "device not paired"
  schemas:
    Error:
      type: object
      description: Standard error response returned by all endpoints on failure.
      required: [status, detail]
      properties:
        status:
          type: integer
          description: HTTP status code
          example: 400
        detail:
          type: string
          description: Human-readable error message
          example: "invalid request body"

    # ─── Instance Schemas ──────────────────────────────────────────
...
    MarkAsReadRequest:
      type: object
      required: [chatId, senderId, receiptType]
      properties:
        chatId:
          type: string
          description: Chat JID
        senderId:
          type: string
          description: Sender JID
        receiptType:
          type: string
          enum: [delivered, sender, read, played]
...
    ChatListItem:
      type: object
      properties:
        id:
          type: string
          description: Chat JID
          example: "1234567890@s.whatsapp.net"
        lid:
          type: string
          description: LID-based JID for the chat (omitted for groups or if unknown)
          example: "222424242@lid"
        isGroup:
          type: boolean
        isArchived:
          type: boolean
        isPinned:
          type: boolean
        isMuted:
          type: boolean
        muteEndTime:
          type: string
          format: date-time
          description: When the mute expires (omitted if not muted)
        pushName:
          type: string
          description: Contact push name (omitted if unknown)
        businessName:
          type: string
          description: Business name (omitted if not a business)
        fullName:
          type: string
          description: Full name from contacts (omitted if unknown)
        lastActivity:
          type: string
          format: date-time
          description: Timestamp of last activity (omitted if unknown)
    ChatPictureResponse:
      type: object
      properties:
        pictureId:
          type: string
        pictureUrl:
          type: string
    BusinessProfileResponse:
      type: object
      properties:
        id:
          type: string
        address:
          type: string
        email:
          type: string
        description:
          type: string
        website:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
        memberSince:
          type: string
        categories:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              name:
                type: string
        businessHours:
          type: array
          items:
            type: object
            properties:
              dayOfWeek:
                type: string
              mode:
                type: string
              openTime:
                type: string
              closeTime:
                type: string
        businessHoursTimeZone:
          type: string
        profileOptions:
          type: object
          additionalProperties:
            type: string
    SetPresenceRequest:
      type: object
      required: [state]
      properties:
        state:
          type: string
          enum: [typing, paused, recording]
    SetChatEphemeralRequest:
      type: object
      required: [expiration]
      properties:
        expiration:
          type: string
          enum: [off, 24h, 7d, 90d]
          description: Disappearing messages timer
    MuteChatRequest:
      type: object
      required: [duration]
      properties:
        duration:
          type: string
          enum: [8h, 1w, always, off]
    PinChatRequest:
      type: object
      properties:
        pinned:
          type: boolean
    ArchiveChatRequest:
      type: object
      properties:
        archived:
          type: boolean
    ChatMarkAsReadRequest:
      type: object
      properties:
        read:
          type: boolean
    RequestMessagesRequest:
      type: object
      required: [lastMessageId, lastMessageSenderId]
      properties:
        lastMessageId:
          type: string
          description: ID of the last known message in the chat
        lastMessageSenderId:
          type: string
          description: Phone number or JID of the sender of the last known message
        count:
          type: integer
          description: Number of messages to request (default 50, max 500)
          minimum: 1
          maximum: 500
          default: 50

    # ─── Contact Schemas ─────────────────────────────────────────
...
    SetMyPresenceRequest:
      type: object
      required: [presence]
      properties:
        presence:
          type: string
          enum: [available, unavailable]
          description: Presence state

    # ─── Call Schemas ──────────────────────────────────────────────

    # ─── Identity Schemas ────────────────────────────────────────
...